# cn_project/openapi.json
{"openapi":"3.1.0","info":{"title":"NFC Bus Backend","version":"0.1.0"},"paths":{"/students":{"post":{"summary":"Create Student","operationId":"create_student_students_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/students/me":{"get":{"summary":"Get Current Student","operationId":"get_current_student_students_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/students/me/fcm_token":{"patch":{"summary":"Update Student Fcm Token","operationId":"update_student_fcm_token_students_me_fcm_token_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FcmTokenUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/students/me/nfc":{"patch":{"summary":"Register Nfc","operationId":"register_nfc_students_me_nfc_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NfcUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/students/me/block_nfc":{"patch":{"summary":"Block Nfc","operationId":"block_nfc_students_me_block_nfc_patch","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/students/me/tap_notifications/{notification_id}/respond":{"post":{"summary":"Respond To Tap Notification","operationId":"respond_to_tap_notification_students_me_tap_notifications__notification_id__respond_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"notification_id","in":"path","required":true,"schema":{"type":"string","title":"Notification Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TapNotificationResponse"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/students/me/pending_notifications":{"get":{"summary":"Get Pending Notifications","operationId":"get_pending_notifications_students_me_pending_notifications_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PendingNotification"},"type":"array","title":"Response Get Pending Notifications Students Me Pending Notifications Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/students/me/notifications/{notification_id}/dismiss":{"post":{"summary":"Dismiss Expired Warning","operationId":"dismiss_expired_warning_students_me_notifications__notification_id__dismiss_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"notification_id","in":"path","required":true,"schema":{"type":"string","title":"Notification Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/students/me/trips":{"get":{"summary":"Get Student Trips","operationId":"get_student_trips_students_me_trips_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Trip"},"type":"array","title":"Response Get Student Trips Students Me Trips Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/students/{student_id}":{"delete":{"summary":"Delete Student","operationId":"delete_student_students__student_id__delete","parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"string","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drivers":{"post":{"summary":"Create Driver","operationId":"create_driver_drivers_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusDriverCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusDriver"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drivers/me":{"get":{"summary":"Get Current Driver","operationId":"get_current_driver_drivers_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusDriver"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/drivers/me/fcm_token":{"patch":{"summary":"Update Driver Fcm Token","operationId":"update_driver_fcm_token_drivers_me_fcm_token_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FcmTokenUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/drivers/me/trips":{"get":{"summary":"Get Driver Trips","operationId":"get_driver_trips_drivers_me_trips_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Trip"},"type":"array","title":"Response Get Driver Trips Drivers Me Trips Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/drivers/me/trips/{trip_id}/details":{"get":{"summary":"Get Driver Trip Details","operationId":"get_driver_trip_details_drivers_me_trips__trip_id__details_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"string","title":"Trip Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drivers/{driver_id}":{"delete":{"summary":"Delete Driver","operationId":"delete_driver_drivers__driver_id__delete","parameters":[{"name":"driver_id","in":"path","required":true,"schema":{"type":"string","title":"Driver Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drivers/me/trips/active":{"get":{"summary":"Get Active Trip","operationId":"get_active_trip_drivers_me_trips_active_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"summary":"Create And Start Active Trip","operationId":"create_and_start_active_trip_drivers_me_trips_active_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/drivers/me/trips/active/end":{"patch":{"summary":"End Active Trip","operationId":"end_active_trip_drivers_me_trips_active_end_patch","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/drivers/me/trips/active/tap":{"post":{"summary":"Tap Nfc In Active Trip","operationId":"tap_nfc_in_active_trip_drivers_me_trips_active_tap_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/trips":{"post":{"summary":"Create Trip","operationId":"create_trip_trips_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}/start":{"patch":{"summary":"Start Trip","operationId":"start_trip_trips__trip_id__start_patch","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"string","title":"Trip Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}/stop":{"patch":{"summary":"Stop Trip","operationId":"stop_trip_trips__trip_id__stop_patch","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"string","title":"Trip Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}/details":{"get":{"summary":"Get Trip Details","operationId":"get_trip_details_trips__trip_id__details_get","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"string","title":"Trip Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TripDetail"},"title":"Response Get Trip Details Trips  Trip Id  Details Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trip-details":{"post":{"summary":"Add Trip Detail","operationId":"add_trip_detail_trip_details_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripDetailCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/token":{"post":{"summary":"Login","operationId":"login_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_token_post"},"BusDriver":{"properties":{"name":{"type":"string","title":"Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"fcm_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fcm Token"}},"type":"object","required":["name","id","created_at"],"title":"BusDriver"},"BusDriverCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","password"],"title":"BusDriverCreate"},"FcmTokenUpdate":{"properties":{"fcm_token":{"type":"string","title":"Fcm Token"}},"type":"object","required":["fcm_token"],"title":"FcmTokenUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NfcUpdate":{"properties":{"nfc_id":{"type":"string","title":"Nfc Id"}},"type":"object","required":["nfc_id"],"title":"NfcUpdate"},"PendingNotification":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"nfc_id":{"type":"string","title":"Nfc Id"},"trip_id":{"type":"string","format":"uuid","title":"Trip Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"expires_at":{"type":"string","format":"date-time","title":"Expires At"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","nfc_id","trip_id","created_at","expires_at","status"],"title":"PendingNotification"},"Student":{"properties":{"name":{"type":"string","title":"Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"nfc_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nfc Id"},"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"is_nfc_blocked":{"type":"boolean","title":"Is Nfc Blocked","default":false},"fcm_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fcm Token"}},"type":"object","required":["name","id","created_at"],"title":"Student"},"StudentCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"nfc_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nfc Id"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","password"],"title":"StudentCreate"},"TapNotificationResponse":{"properties":{"notification_id":{"type":"string","format":"uuid","title":"Notification Id"},"accepted":{"type":"boolean","title":"Accepted"}},"type":"object","required":["notification_id","accepted"],"title":"TapNotificationResponse"},"Trip":{"properties":{"driver_id":{"type":"string","format":"uuid","title":"Driver Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name","default":"Trip"},"id":{"type":"string","format":"uuid","title":"Id"},"start_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"},"end_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"},"total_students":{"type":"integer","title":"Total Students"},"total_revenue":{"type":"number","title":"Total Revenue"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["driver_id","id","total_students","total_revenue","created_at"],"title":"Trip"},"TripCreate":{"properties":{"driver_id":{"type":"string","format":"uuid","title":"Driver Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name","default":"Trip"}},"type":"object","required":["driver_id"],"title":"TripCreate"},"TripDetail":{"properties":{"trip_id":{"type":"string","format":"uuid","title":"Trip Id"},"nfc_id":{"type":"string","title":"Nfc Id"},"id":{"type":"string","format":"uuid","title":"Id"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"status":{"type":"string","title":"Status","default":"pending"}},"type":"object","required":["trip_id","nfc_id","id","timestamp"],"title":"TripDetail"},"TripDetailCreate":{"properties":{"trip_id":{"type":"string","format":"uuid","title":"Trip Id"},"nfc_id":{"type":"string","title":"Nfc Id"}},"type":"object","required":["trip_id","nfc_id"],"title":"TripDetailCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}